Does store list locally now
